Retrieves essential base vectors

Collects vector data from the BC Data Catelog for a spcified area of interest.

Usage

create_base_vectors(in_aoi, out_path = "00_raw_inputs/vector")

Arguments

in_aoi

sf polygon

out_path

character string that points the output location

Details

This script downloads the relevant spatial data for BEC zones, VRI, TEM, waterbodies and the road network used in stage 1 of PEM processing. Data is downloaded directly from the BC Data Catalogue using the bcdata package.

Datasets retrieved include:

BEC - Biogeoclimatic Ecosystems Classification are used to define and/or select specific "subzones" within a defined study area. Note this can be also pointed to different file where more updated version is available.

Vegetation Resource Inventory (VRI) - This layer includes a variety of vegetation measure, including cutblock age, TEM data and ..... Detailed data standards can be found here.

Freshwater Atlas - The atlas is separated into the different types of waterbodies (lakes, rivers, wetlands, streams, man made, etc.), which requires a seperate download per type. Alternatively a single combined layer can be downloaded by is limited to linear data type. A parameter within the function can be set to "polygon" (the default option) or "linear".

Road network THIS STILL NEEEDS WORK

Road network - In previous works, the raw road network was found to be too detailed. The raw road network is filtered to only keep named roads, and then the up to date Forest Service Road (FSR) layer is downloaded and merged with the filtered road network to produce a really good representation of where roads are actually located on the landscape.

Fire - Fire and fire intensity are used used to assist in identifying areas where a high cost is applied to reduce sampling. This includes: -current -historical polygons. -severity.

Accessing the consolidated cutblock layer using the bcdata package. The FTen layer is also used to identify very recent cutblocks.

Note the detailed Tenure (private ownership) WHSE_CADASTRE.PMBC_PARCEL_FABRIC_POLY_FA_SVW BCGW

Detailed roads layers. In some cases regionally specific roads are available through other collaborative projects (cumulative effects). These are not directly available through the BCGW.

Examples